Created new authentication controller and moved the authentication functions from user and admin controller to this
